+++ src/com/monyetmabuk/rajawali/tutorials/examples/lights/PointLightFragment.java
@@ -52,7 +52,7 @@ public class PointLightFragment extends AExampleFragment {
 			int count = 0;
 
 			for (int i = 0; i < 360; i += 36) {
-				float radians = MathUtil.degreesToRadians(i);
+				double radians = MathUtil.degreesToRadians(i);
 				int color = 0xfed14f;
 				if (count % 3 == 0)
 					color = 0x10a962;
@@ -75,7 +75,7 @@ public class PointLightFragment extends AExampleFragment {
 			count = 0;
 
 			for (int i = 0; i < 360; i += 12) {
-				float radians = MathUtil.degreesToRadians(i);
+				double radians = MathUtil.degreesToRadians(i);
 				int color = 0xfed14f;
 				if (count % 3 == 0)
 					color = 0x10a962;
@@ -84,8 +84,8 @@ public class PointLightFragment extends AExampleFragment {
 				count++;
 
 				BaseObject3D sphere = rootSphere.clone(false);
-				sphere.setPosition((float) Math.sin(radians) * radius, 0,
-						(float) Math.cos(radians) * radius);
+				sphere.setPosition(Math.sin(radians) * radius, 0,
+						Math.cos(radians) * radius);
 				sphere.setMaterial(sphereMaterial);
 				sphere.setColor(color);
 				sphere.addLight(pointLight);
